requirements.txt
streamlit==1.46.0
google-generativeai==0.5.0
python-dotenv==1.0.1
playsound==1.3.0
pydub==0.25.1
